\name{assign_polygons}
\alias{assign_polygons}
\docType{package}
\title{
Assign the polygons in the original spatial data to their new location.}
\description{Assigns each polygon in the original file to a new location in the gridded geometry. Enables the use of three assignment algorithms. Returns a SpatialPolygonsDataFrame.}
\usage{
assign_polygons()
}
\examples{
\dontrun{
assign_polygons()
}
}
