\name{calculate_cell_size}
\alias{calculate_cell_size}
\docType{package}
\title{
Calculate size of grid items.
}
\description{
Given an input multipolgyon spatial data frame this function calculates the required cell size of a regular or hexagonal grid.}
\usage{
calculate_cell_size()
}
\examples{
\dontrun{
calculate_cell_size()
}
}
