\name{get_shape_details}
\alias{get_shape_details}
\docType{package}
\title{
Extract details from provided polygons.
}
\description{
Extract spatial extent, range and other geospatial features from the output of read_polygons. Items are returned as a list for use in calculate_cell_size.}
\usage{
get_shape_details()
}
\examples{
\dontrun{
get_shape_details()
}
}
