\name{read_polygons}
\alias{read_polygons}
\docType{package}
\title{
Import spatial data.
}
\description{
Simple function to read spatial data into a SpatialPolygonsDataFrame. Based on st_read from package sf.
}
\usage{
read_polygons()
See st_read in package sf.
}
\examples{
\dontrun{
read_polygons()
}
}
