% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_contiguous.R
\name{sample_contiguous}
\alias{sample_contiguous}
\title{Sample Contiguous}
\usage{
sample_contiguous(shape, n = 5)
}
\arguments{
\item{shape}{sf shape object}

\item{n}{number of shapes to sample}
}
\value{
indices of a contiguous sample
}
\description{
Sample Contiguous
}
\examples{
library(sf)
input_file <- system.file("extdata", "london_LA.json", package = "geogrid")
original_shapes <- st_read(input_file) \%>\% st_set_crs(27700)
sample_shapes <- sample_contiguous(original_shapes, n = 5)
}
