% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shape_details.R
\name{get_shape_details}
\alias{get_shape_details}
\title{Extract details from provided polygons (deprecated).}
\usage{
get_shape_details(input_shape)
}
\arguments{
\item{input_shape}{A "SpatialPolygonsDataFrame" object representing the
original spatial polygons.}
}
\description{
Extract spatial extent, range and other geospatial features from the output
of read_polygons. Items are returned as a list for use in
\code{\link{calculate_grid}}.
}
