% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_polygons.R
\name{read_polygons}
\alias{read_polygons}
\title{Import spatial data.}
\usage{
read_polygons(file)
}
\arguments{
\item{file}{A file path pointing to a shapefile or GeoJSON file, or a character string holding GeoJSON data. See the \code{dsn} argument of \code{\link[sf]{st_read}} for more details.}
}
\description{
Simple function to read spatial data into a SpatialPolygonsDataFrame. Based on st_read from package sf.
}
